Finalized the design and adjusted data sets

+++ api/DTOs/BudgetAllDTO.cs
@@ -9,7 +9,7 @@ namespace api.DTOs
     public class BudgetAllDTO
     {
         public int Id { get; set; }
-        public int Income { get; set; }
+        public float Income { get; set; }
         public string Frequency { get; set; }
         public int Month { get; set; }
         public int Year { get; set; }
